DirectedKeys

Documentation

encodeKeyRaw :: (Serialize key, Serialize source, Serialize destination, Serialize datetime) => key -> source -> destination -> datetime -> DirectedKeyRaw key source destination datetime

encodeKey :: (Serialize key, Serialize source, Serialize destination, Serialize datetime) => DirectedKeyRaw key source destination datetime -> ByteString

encodeKeyPart :: (Serialize key, Serialize source, Serialize destination, Serialize datetime) => DirectedKeyRaw key source destination datetime -> ByteString

decodeKey :: (Serialize key, Serialize source, Serialize destination, Serialize datetime) => ByteString -> Either String (DirectedKeyRaw key source destination datetime)

decodeKeyPart :: (Serialize key, Serialize source, Serialize destination, Serialize datetime) => ByteString -> Either String (DirectedKeyRaw key source destination datetime)

toEscapedCharacters :: Map Char ByteString

parseFilename :: ByteString -> ByteString

Use this function to create a filename from a bytestring, it escapes all invalid unix filename characters

decodeFilename :: ByteString -> ByteString

decodeFilename returns a Bytestring that has unescaped all the characters for filenames